FinalizaciónPráctica

+++ src/main_rpn_t.cc
@@ -1,20 +1,23 @@
-// AUTOR: 
-// FECHA: 
-// EMAIL: 
-// VERSION: 2.0
-// ASIGNATURA: Algoritmos y Estructuras de Datos
-// PRÁCTICA Nº: 5
-// ESTILO: Google C++ Style Guide
-// COMENTARIOS: Programa principal para probar la calculadora de expresiones 
-//              en notaión polaca inversa (RPN, Reverse Polish Notation)
-// COMPILACIÓN: g++ main_rpn_t.cc -o main_rpn_t
-// EJECUCIÓN:
-//   ./main_rpn_t < data_rpn_t_1.txt
-//   ./main_rpn_t < data_rpn_t_2.txt
-//   ./main_rpn_t < data_rpn_t_3.txt
-
-// Compilar con:
-// 
+/**
+  * Universidad de La Laguna
+  * Escuela Superior de Ingeniería y Tecnología
+  * Grado en Ingeniería Informática
+  * Informática Básica
+  *
+  * @file queue_l_t.h
+  * @author Jose Dorta Luis alu0101414676@ull.edu.es
+  * @date 24 april 2021
+  * @brief programa proporcionado por el profesorado para realizar la práctica 5 de Algoritmos y estructura de datos
+  * @bug ningún bug conocido
+  * Compilación:
+  * g++ -g main_rpn_t.cc -o main_rpn_t
+  * 
+  * Ejecución:
+  * ./main_rpn_t < data_rpn_t_1.txt  
+  * ./main_rpn_t < data_rpn_t_3.txt
+  * ./main_rpn_t < data_rpn_t_2.txt
+  * 
+  */
 
 #include <iostream>
 
@@ -25,7 +28,7 @@
 using namespace std;
 
 int main(void) {
-  rpn_t<stack_l_t<int>> calculadora;
+  rpn_t<stack_l_t<double>> calculadora;
   queue_l_t<char> cola;
 	
   while (!cin.eof()) 	{
@@ -40,7 +43,7 @@ int main(void) {
   cout << "Expresión a evaluar: ";	
   cola.write();
 
-  int r = calculadora.evaluate(cola);
+  double r = calculadora.evaluate(cola);
   cout << "Resultado: " << r << endl;
 
   return 0;
